hamburger menu fixed and sidebar html done
have added the sidebar htmla from the list of items from the netmatters website. This needs to be styled, hidden and postitioned.

Have also redesigned the hamburger menu whist also adding logic to hid e a span and rotate the other two, using a pseudo class - nth child to traverse the DOM.