mobile break points
added styling to the header, and the mobile breakpoint styles. Ensuring it matches close to the netmatters style.
Added the pgone icon the html.
Tidied up the div tags in the index.html for the header acions sections, ensuring opening and closing divs line up.
Update of brekpoints for smal to md -1px. Ensuring styling is close to netmatters